+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpResponseInterceptorUnitCallback.java
@@ -30,8 +30,6 @@ import io.ballerina.stdlib.http.api.nativeimpl.ModuleUtils;
 import io.ballerina.stdlib.http.api.nativeimpl.connection.Respond;
 import io.ballerina.stdlib.http.transport.message.HttpCarbonMessage;
 
-import java.util.Objects;
-
 /**
  * {@code HttpResponseInterceptorUnitCallback} is the responsible for acting on notifications received from Ballerina
  * side when a response interceptor service is invoked.
@@ -70,7 +68,9 @@ public class HttpResponseInterceptorUnitCallback extends HttpCallableUnitCallbac
 
     @Override
     public void notifyFailure(BError error) { // handles panic and check_panic
-        invokeErrorInterceptors(error, false);
+        cleanupRequestMessage();
+        sendFailureResponse(error);
+        System.exit(1);
     }
 
     public void invokeErrorInterceptors(BError error, boolean printError) {
@@ -79,10 +79,6 @@ public class HttpResponseInterceptorUnitCallback extends HttpCallableUnitCallbac
             error.printStackTrace();
         }
         returnErrorResponse(error);
-    }
-
-    public void sendFailureResponse(BError error) {
-        HttpUtil.handleFailure(requestMessage, error, false);
     }
 
     private void printStacktraceIfError(Object result) {
@@ -169,10 +165,6 @@ public class HttpResponseInterceptorUnitCallback extends HttpCallableUnitCallbac
                 stopObserverContext();
                 dataContext.notifyOutboundResponseStatus(null);
                 printStacktraceIfError(result);
-                Object isPanic = requestMessage.getProperty(HttpConstants.INTERCEPTOR_SERVICE_PANIC_ERROR);
-                if (Objects.nonNull(isPanic) && (boolean) isPanic) {
-                    System.exit(1);
-                }
             }
 
             @Override